\name{expansions-package}
\alias{expansions-package}
\alias{expansions}
\docType{package}
\title{
  Expansions of integer and real numbers
}
\description{
  Expansions of integer and real numbers in an integer base, odometers and adic addition
}
\details{
  See the examples.
}
\author{
Stéphane Laurent.

Maintainer: Stéphane Laurent <laurent_step@yahoo.fr>
}
\keyword{ package }
