% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floatExpand.R
\name{floatExpand}
\alias{floatExpand}
\title{Float expansion}
\usage{
floatExpand(x, base = 2)
}
\arguments{
\item{x}{a nonnegative number}

\item{base}{the base of the expansion (integer)}
}
\value{
A list representing the scientific notation of the expansion; the digits in the first component and the exponent in the second one.
}
\description{
Expansion of a positive real number in a given integer base.
}
\examples{
floatExpand(1.125)
(1*1/2 + 0*1/2^2 + 1*1/2^3) * 2^1
}

