% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floatExpand.R
\name{floatExpand01}
\alias{floatExpand01}
\title{Float expansion between 0 and 1}
\usage{
floatExpand01(u, base = 2)
}
\arguments{
\item{u}{real number between 0 and 1}

\item{base}{the base of the expansion (integer)}
}
\value{
The digits of the expansion of \code{u} in a vector.
}
\description{
Expansion of a number between 0 and 1 in a given integer base.
}
\examples{
floatExpand01(0.125)
0*1/2 + 0*1/2^2 + 1*1/2^3
floatExpand01(1/3+1/3^3, base=3)
}

