% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intExpand.R
\name{intAtBase}
\alias{intAtBase}
\title{Integer expansion}
\usage{
intAtBase(n, base = 2L)
}
\arguments{
\item{n}{the integer to be expanded}

\item{base}{the base of the expansion}
}
\value{
the expansion of \code{n} in a vector
}
\description{
Expansion of a positive integer in a given base
}
\examples{
intAtBase(41, base=3)
2*1 + 1*3 + 1*3^2 + 1*3^3
floatExpand(41, base=3)
(1*1/3 + 1*1/3^2 + 1*1/3^3 + 2*1/3^4) * 3^4
}

