% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intExpand.R
\name{intAtBaseFib}
\alias{intAtBaseFib}
\title{Integer expansion of a Fibonacci number}
\usage{
intAtBaseFib(n, base = 2L)
}
\arguments{
\item{n}{the index of the Fibonacci number to be expanded}

\item{base}{the base of the expansion}
}
\value{
The expansion of the \code{n}-th Fibonacci number in a vector.
}
\description{
Expansion of a Fibonacci number in an integer base
}
\examples{
intAtBaseFib(6, base=3) == intAtBase(8, base=3)
}

