% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intExpand.R
\name{intAtBasePower}
\alias{intAtBasePower}
\title{Integer expansion of a power}
\usage{
intAtBasePower(m, p, base = 2L)
}
\arguments{
\item{m}{the integer to be exponentiated}

\item{p}{the integer exponent}

\item{base}{the base of the expansion}
}
\value{
The expansion of \code{m^p} in a vector.
}
\description{
Expansion of \code{m^p} in an integer base
}
\examples{
intAtBasePower(2, 5, base=3) == intAtBase(2^5, base=3)
}

