% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intToAry.R
\name{intToAry}
\alias{intToAry}
\title{Integer to ary-expansion}
\usage{
intToAry(n, sizes)
}
\arguments{
\item{n}{integer to be expanded}

\item{sizes}{the base of the expansion}
}
\value{
A vector of length \code{prod(sizes)}, the expansion of \code{n}.
}
\description{
Ary-expansion of an integer.
}

