% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odometers.R
\name{odometer}
\alias{odometer}
\alias{odometer_iterated}
\title{Odometer}
\usage{
odometer(x, base = 2L)

odometer_iterated(x, n, base = 2L)
}
\arguments{
\item{x}{a sequence of digits in the given base}

\item{base}{integer, the base of the expansion}

\item{n}{integer, power of the iteration}
}
\value{
\code{odometer} returns the transformation of \code{x} by the odometer; \code{odometer_iterated} returns the \code{n}-th iterate of the odometer.
}
\description{
The odometer in a given integer base
}
\examples{
odometer(c(0,1))
odometer(c(1,1))
odometer_iterated(c(0,1), n=2)
odometer_iterated(0, n=13) == intAtBase(13)
}

