% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odometers.R
\name{sumadic}
\alias{sumadic}
\title{Sum of two adic integers}
\usage{
sumadic(x1, x2, base = 2L)
}
\description{
Sum of two adic integers in a given base
}
\examples{
sumadic(c(0,1), c(1,1))
sumadic(c(0,1), 1)
odometer(c(0,1))
}

