% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floatExpand.R
\name{floatExpand01}
\alias{floatExpand01}
\title{Float expansion between 0 and 1}
\usage{
floatExpand01(x, base = 2)
}
\arguments{
\item{x}{the number to be expanded}

\item{base}{the base of the expansion (integer)}
}
\value{
The digits of the expansion in a vector.
}
\description{
Expansion of a number between 0 and 1 in a given integer base.
}
\examples{
floatExpand01(0.125)
floatExpand01(1/3+1/27, base=3)
}

