% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intToAry.R
\name{intToAry}
\alias{intToAry}
\title{Integer to ary-expansion}
\usage{
intToAry(n, sizes)
}
\arguments{
\item{n}{integer to be expanded}

\item{sizes}{the base of the expansion}
}
\value{
A vector of length \code{prod(sizes)}, the expansion of \code{n}.
}
\description{
Ary-expansion of an integer.
}
\examples{
intToAry(77, c(3,4,7))
2*1 + 1*3 + 6*(3*4)
# Cartesian product {0,1}x{0,1}:
sapply(0:3, function(x) intToAry(x, sizes=c(2,2)))
}

